Formerly ../hurd/hurdinit.c.~11~

+++ hurd/hurdinit.c
@@ -24,7 +24,7 @@ Cambridge, MA 02139, USA.  */
 
 
 struct _hurd_port *_hurd_ports;
-size_t _hurd_nports;
+unsigned int _hurd_nports;
 mode_t _hurd_umask;
 
 void _hurd_proc_init (char **argv);
